+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: mbalon-s <mbalon-s@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/01 18:14:51 by mbalon-s          #+#    #+#             */
-/*   Updated: 2019/01/31 16:44:04 by mbalon-s         ###   ########.fr       */
+/*   Updated: 2019/01/31 16:50:46 by mbalon-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -154,7 +154,7 @@ void				ft_fmatrixfree(t_fmatrix *matrix);
 t_imatrix			ft_fmatrix2imatrix(t_fmatrix matrix);
 t_fmatrix			ft_imatrix2fmatrix(t_imatrix matrix);
 /* hashmap */
-unsigned int		ft_murmurhash(const void *key, size_t len,
+unsigned int		ft_murmurhash(const void *key, unsigned int len,
 									unsigned int seed);
 // get
 // set
